Optionally get chrome nav config
In this patch we optionally grab the chrome nav config from cloud-services-config if the INCLUDE_CHROME_CONFIG envvar is set to true in the build script that sourced this script. My original intention was to do this in frontend-build-container but I can't git clone there easily. I realized this script runs in jenkins with git ssh creds from vault, so I should be able to do this here.

For all apps but chrome this will be a noop. We will then copy the config into the container via the Dockerfile in the app repo.